+++ gestionart_frontend_ruben_y_jessica/lib/data/repositories/SolicitudExclusivaRepository.dart
@@ -1,4 +1,4 @@
-import 'dart:ffi';
+
 
 import 'package:gestionart_frontend_ruben_y_jessica/data/enums/EstadoSolicitud.dart';
 import 'package:gestionart_frontend_ruben_y_jessica/data/models/SolicitudExclusiva.dart';
@@ -8,7 +8,7 @@ class Solicitudexclusivarepository {
   ApiService _apiService;
   Solicitudexclusivarepository(ApiService? apiService) : _apiService = apiService ?? ApiService();
 
-  Future<void> crearSolicitudExclusiva(Long idComprador, Long idArticulo, String mensaje, Long idVendedor) async {
+  Future<void> crearSolicitudExclusiva(int idComprador, int idArticulo, String mensaje, int idVendedor) async {
     try {
       final response = await _apiService.dio.post("/solicitudes-exclusivas", data: {
         "idComprador": idComprador,
@@ -24,7 +24,7 @@ class Solicitudexclusivarepository {
     }
   }
 
-  Future<void> obtenerPorId(Long id) async {
+  Future<void> obtenerPorId(int id) async {
     try {
       final response = await _apiService.dio.get("/solicitudes-exclusivas/$id");
       if (response.statusCode != 200) {
@@ -35,7 +35,7 @@ class Solicitudexclusivarepository {
     }
   }
 
-  Future<List<Solicitudexclusiva>> obtenerPorVendedor(Long idVendedor) async {
+  Future<List<Solicitudexclusiva>> obtenerPorVendedor(int idVendedor) async {
     try {
       final response = await _apiService.dio.get("/solicitudes-exclusivas/vendedor/$idVendedor");
       if (response.statusCode == 200) {
@@ -52,7 +52,7 @@ class Solicitudexclusivarepository {
     }
   }
 
-  Future<List<Solicitudexclusiva>> obtenerPorVendedorYEstado(Long idVendedor, Estadosolicitud estado) async {
+  Future<List<Solicitudexclusiva>> obtenerPorVendedorYEstado(int idVendedor, Estadosolicitud estado) async {
     try {
       final response = await _apiService.dio.get("/solicitudes-exclusivas/vendedor/$idVendedor/estado/$estado");
       if (response.statusCode == 200) {
@@ -69,7 +69,7 @@ class Solicitudexclusivarepository {
     }
   }
 
-  Future<List<Solicitudexclusiva>> obtenerPorComprador(Long idComprador) async {
+  Future<List<Solicitudexclusiva>> obtenerPorComprador(int idComprador) async {
     try {
       final response = await _apiService.dio.get("/solicitudes-exclusivas/comprador/$idComprador");
       if (response.statusCode == 200) {
@@ -86,7 +86,7 @@ class Solicitudexclusivarepository {
     }
   }
   
-  Future<List<Solicitudexclusiva>> obtenerPorCompradorYEstado(Long idComprador, Estadosolicitud estado) async {
+  Future<List<Solicitudexclusiva>> obtenerPorCompradorYEstado(int idComprador, Estadosolicitud estado) async {
     try {
       final response = await _apiService.dio.get("/solicitudes-exclusivas/comprador/$idComprador/estado/$estado");
       if (response.statusCode == 200) {
@@ -120,7 +120,7 @@ class Solicitudexclusivarepository {
     }
   } 
 
-  Future<void> actualizarEstado(Long id, Estadosolicitud nuevoEstado) async {
+  Future<void> actualizarEstado(int id, Estadosolicitud nuevoEstado) async {
     try {
       final response = await _apiService.dio.put("/solicitudes-exclusivas/$id/estado/$nuevoEstado");
       if (response.statusCode != 200) {
