+++ gestionart_frontend_ruben_y_jessica/lib/data/repositories/Usuariorepository.dart
@@ -1,4 +1,4 @@
-import 'dart:ffi';
+
 
 import 'package:gestionart_frontend_ruben_y_jessica/data/models/Usuario.dart';
 import 'package:gestionart_frontend_ruben_y_jessica/data/services/ApiService.dart';
@@ -9,7 +9,7 @@ class UsuarioRepository {
 
   UsuarioRepository(ApiService? apiService) : _apiService = apiService ?? ApiService();
 
-  Future<Usuario> getUsuarioPorId(Long id) async {
+  Future<Usuario> getUsuarioPorId(int id) async {
 
     try{
       final response = await _apiService.dio.get("/users/$id");
