+++ gestionart_frontend_ruben_y_jessica/lib/data/repositories/VendedorRepository.dart
@@ -1,4 +1,4 @@
-import 'dart:ffi';
+
 
 import 'package:gestionart_frontend_ruben_y_jessica/data/models/Vendedor.dart';
 import 'package:gestionart_frontend_ruben_y_jessica/data/services/ApiService.dart';
@@ -8,7 +8,7 @@ class Vendedorrepository {
   final ApiService apiService;
   Vendedorrepository(ApiService? apiService) : apiService = apiService ?? ApiService();
 
-  Future<void> actualizarVendedor(Long id, String correoElectronico, String nombre, String descripcionPerfil, String imagen, String contrasena) async {
+  Future<void> actualizarVendedor(int id, String correoElectronico, String nombre, String descripcionPerfil, String imagen, String contrasena) async {
     try {
       final response = await apiService.dio.put("/vendedores/$id", data: {
         "correoElectronico": correoElectronico,
@@ -41,7 +41,7 @@ class Vendedorrepository {
       throw Exception("Error al obtener los vendedores: $e");
     }
   }
-  Future<Vendedor> getVendedorPorId(Long id) async {
+  Future<Vendedor> getVendedorPorId(int id) async {
     try {
       final response = await apiService.dio.get("/vendedores/$id");
       if (response.statusCode == 200) {
@@ -80,7 +80,7 @@ class Vendedorrepository {
     }
   }
 
-  Future<void> eliminarVendedor(Long id) async {
+  Future<void> eliminarVendedor(int id) async {
     try {
       final response = await apiService.dio.delete("/vendedores/$id");
       if (response.statusCode != 200) {
